Seperated api repo

+++ Recommendations/Models/CatalogItemMandatory.cs
@@ -5,7 +5,7 @@
         public string Id { get; set; }
         public string Title { get; set; }
         public string Type { get; set; }
-        public double RecommendationRating { get; set; }
+        public decimal RecommendationRating { get; set; }
 
         public CatalogItem()
         {
